Updated th UI of Fuel Registration & update Forms a nd added custom validations.

+++ src/pages/fuel/Fuel.jsx
@@ -91,11 +91,14 @@ const Fuel = () => {
   };
 
   const handleDelete = async (id) => {
-    try {
-      await deleteDoc(doc(db, "fuel", id));
-      setData(data.filter((item) => item.id !== id));
-    } catch (error) {
-      console.log(error);
+    const confirmed = confirm("Are you sure you want to delete this fuel?");
+    if (confirmed) {
+      try {
+        await deleteDoc(doc(db, "fuel", id));
+        setData(data.filter((item) => item.id !== id));
+      } catch (error) {
+        console.log(error);
+      }
     }
   };
